refactor: Introduce `configuration/resolve-variables` as separate module
