+++ src/index.js
@@ -16,7 +16,7 @@ const sendTelemetry = require('./utils/telemetry/send');
 const ServerlessError = require('./serverless-error');
 const handleError = require('./handle-error');
 const colors = require('./cli/colors');
-const { resolveConfigurationVariables } = require('./configuration/configuration');
+const resolveConfigurationVariables = require('./configuration/resolve-variables');
 const resolveConfigurationPath = require('./configuration/resolve-path');
 const readConfiguration = require('./configuration/read');
 const validateConfiguration = require('./configuration/validate');
